requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.29
pydantic==2.6.4
requests==2.31.0
chromadb==0.4.24
sentence-transformers==2.6.1
streamlit==1.32.2
fpdf2==2.7.9
